// src/generated/prisma/enums.ts
export const AiUsageStep = {
  MODERATION: 'MODERATION',
  CHAT_COMPLETION: 'CHAT_COMPLETION',
  REPHRASING: 'REPHRASING',
  EMBEDDINGS: 'EMBEDDINGS',
  RERANKING: 'RERANKING'
} as const

export type AiUsageStep = (typeof AiUsageStep)[keyof typeof AiUsageStep]